extract utility function to fetch resources from off-chain storage
